refactoring directories
