+++ src/app/modules/cards-collection/components/list/cards-collection-list.component.html
