+++ src/app/modules/cards-collection/components/list/index.ts
@@ -1,0 +1,3 @@
+export { CardsCollectionListComponent } from './cards-collection-list.component';
+export { CardsCollectionListItemComponent } from './item/cards-collection-list-item.component';
+export { MyCardsListFilterComponent } from './filter/my-cards-list-filter.component';
