+++ src/app/modules/cards-collection/components/list/item/cards-collection-list-item.component.ts
