--- src/app/modules/cards-collection/list/index.ts
@@ -1,3 +1,0 @@
-export { CardsCollectionListComponent } from './cards-collection-list.component';
-export { CardsCollectionListItemComponent } from './components/item/cards-collection-list-item.component';
-export { MyCardsListFilterComponent } from './components/filter/my-cards-list-filter.component';
